♻️ refactor(build): Refactor QuoterService with platform abstraction
Replace hard-coded POSIX daemon management with platform-aware
implementation that supports both Windows and POSIX.

Changes:
- Use QuoterPlatform interface for platform-specific daemon operations
- Add elixirService parameter (for future use)
- Store process reference for Windows cleanup
- Platform detection and logging for debugging

Windows implementation uses ProcessBuilder to manage the daemon as a
foreground process, while POSIX continues to use the native daemon
command. Both platforms share the same RPC-based status checking.
